# postman/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API Raizes do Nordeste",
        "description": "Back-end da rede de lanchonetes Raizes do Nordeste.\n\nAtende varias unidades da franquia e varios canais de venda\n(APP, TOTEM, BALCAO, PICKUP e WEB).\n\nComo testar:\n1. Chame POST /auth/login com um usuario do seed.\n2. Clique em Authorize aqui em cima e cole o accessToken.\n3. As rotas protegidas passam a funcionar.\n\nUsuarios do seed (senha Senha@123):\n- admin@raizes.com.br (ADMIN)\n- gerente.recife@raizes.com.br (GERENTE)\n- atendente.recife@raizes.com.br (ATENDENTE)\n- cozinha.recife@raizes.com.br (COZINHA)\n- cliente@exemplo.com (CLIENTE)",
        "contact": {
            "name": "Gean Brandao"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local"
        },
        {
            "url": "/",
            "description": "Mesma origem que serviu esta pagina"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
    "paths": {
        "/unidades/{unidadeId}": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Detalhar unidade",
                "description": "Dados de uma unidade pelo id. Rota publica: o cliente precisa saber onde a loja fica antes de pedir.",
                "operationId": "detalhar",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade n\u00e3o encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "put": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Atualizar unidade",
                "description": "Exclusivo do perfil ADMIN.",
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnidadeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnidadeResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade n\u00e3o encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/produtos/{produtoId}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Detalhar produto",
                "description": "Dados de um produto do catalogo da rede. Traz tambem o que foi inativado, com ativo = false, porque pedido antigo aponta para ele.",
                "operationId": "detalhar_1",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdutoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "description": "Perfis ADMIN e GERENTE.",
                "operationId": "atualizar_1",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProdutoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Inativar produto",
                "description": "Exclusivo do perfil ADMIN. Marca como inativo em vez de apagar, para n\u00e3o quebrar os pedidos antigos que apontam para o produto.",
                "operationId": "inativar",
                "parameters": [
                    {
                        "name": "produtoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto inativado"
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/pedidos/{pedidoId}/pagamentos": {
            "post": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Solicitar pagamento",
                "description": "Pede a cobran\u00e7a ao gateway externo e registra o que voltou.\n\nO valor cobrado e sempre o total do pedido, calculado pelo servidor.\n\n**Desfechos no gateway simulado**, controlados pelo `tokenPagamento`:\n- `tok_recusa...` \u2192 pagamento RECUSADO, pedido vai para PAGAMENTO_RECUSADO\n- `tok_timeout...` \u2192 gateway n\u00e3o responde, pagamento fica PENDENTE e o pedido n\u00e3o se mexe, aguardando o callback\n- qualquer outro valor, ou nenhum \u2192 pagamento APROVADO, pedido vai para PAGO\n\nMandar a mesma `chaveIdempotencia` duas vezes devolve o resultado que ja existe, em vez de cobrar de novo.",
                "operationId": "solicitar",
                "parameters": [
                    {
                        "name": "pedidoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolicitarPagamentoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Pagamento processado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido n\u00e3o encontrado ou n\u00e3o visivel",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Pedido ja pago, ou cancelado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos invalidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/pagamentos/callback": {
            "post": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Callback do gateway",
                "description": "Resolve um pagamento que ficou pendente porque o gateway n\u00e3o respondeu na hora.\n\nProtegido pelo header `X-Gateway-Assinatura`, n\u00e3o por token de usuario: quem chama e o gateway.\n\nCallback repetido para pagamento ja resolvido e ignorado, porque gateway reenvia webhook quando n\u00e3o recebe confirma\u00e7\u00e3o.",
                "operationId": "callback",
                "parameters": [
                    {
                        "name": "X-Gateway-Assinatura",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CallbackPagamentoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Callback processado ou ignorado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Assinatura ausente ou invalida",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pagamento n\u00e3o encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/pedidos/{pedidoId}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Detalhar pedido",
                "description": "Pedido que n\u00e3o pertence a quem esta consultando devolve 404, n\u00e3o 403: dizer 'existe mas n\u00e3o e seu' ja confirmaria a existencia dele.",
                "operationId": "detalhar_2",
                "parameters": [
                    {
                        "name": "pedidoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido n\u00e3o encontrado ou n\u00e3o visivel a esse perfil",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/pagamentos/{pagamentoId}": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Consultar pagamento",
                "description": "A visibilidade e a mesma do pedido: quem n\u00e3o ve o pedido n\u00e3o ve quanto foi cobrado nele.",
                "operationId": "detalhar_3",
                "parameters": [
                    {
                        "name": "pagamentoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagamentoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pagamento n\u00e3o encontrado ou n\u00e3o visivel",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/auditoria": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Consultar trilha de auditoria",
                "description": "Exclusivo do perfil ADMIN.\n\nA\u00e7\u00f5es rastreadas: cria\u00e7\u00e3o, cancelamento e mudan\u00e7a de status de pedido, movimenta\u00e7\u00e3o de estoque, solicita\u00e7\u00e3o de pagamento, ac\u00famulo e resgate de pontos, registro e revoga\u00e7\u00e3o de consentimento.",
                "operationId": "consultar",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "in": "query",
                        "description": "Filtra por quem fez a a\u00e7\u00e3o",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "entidade",
                        "in": "query",
                        "description": "PEDIDO, ESTOQUE, PAGAMENTO, FIDELIDADE ou CONSENTIMENTO",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros da trilha",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginaResponseLogAuditoriaResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Perfil sem permiss\u00e3o",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UnidadeRequest": {
                "type": "object",
                "description": "Dados de uma unidade",
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Raizes Boa Viagem",
                        "maxLength": 120,
                        "minLength": 3
                    },
                    "cidade": {
                        "type": "string",
                        "example": "Recife",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "uf": {
                        "type": "string",
                        "example": "PE",
                        "minLength": 1,
                        "pattern": "^[A-Za-z]{2}$"
                    },
                    "tipoOperacao": {
                        "type": "string",
                        "description": "COMPLETA prepara o cardapio inteiro; REDUZIDA prepara so parte dele",
                        "enum": [
                            "COMPLETA",
                            "REDUZIDA"
                        ],
                        "example": "COMPLETA"
                    }
                },
                "required": [
                    "cidade",
                    "nome",
                    "tipoOperacao",
                    "uf"
                ]
            },
...
            "ErrorResponse": {
                "type": "object",
                "description": "Formato padr\u00e3o de erro da API",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "ESTOQUE_INSUFICIENTE"
                    },
                    "message": {
                        "type": "string",
                        "example": "N\u00e3o ha quantidade suficiente para um ou mais itens."
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-05T12:00:00Z"
                    },
                    "path": {
                        "type": "string",
                        "example": "/pedidos"
                    },
                    "requestId": {
                        "type": "string",
                        "example": "3f1a2b4c-5d6e-7f80-91a2-b3c4d5e6f708"
                    }
                },
                "required": [
                    "details",
                    "error",
                    "message",
                    "timestamp"
                ]
            },
            "UnidadeResponse": {
                "type": "object",
                "description": "Unidade da rede",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "cidade": {
                        "type": "string"
                    },
                    "uf": {
                        "type": "string"
                    },
                    "tipoOperacao": {
                        "type": "string",
                        "enum": [
                            "COMPLETA",
                            "REDUZIDA"
                        ]
                    },
                    "ativa": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "ativa",
                    "cidade",
                    "id",
                    "nome",
                    "tipoOperacao",
                    "uf"
                ]
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Cole aqui o accessToken devolvido por POST /auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}